etl/requirements.txt
apache-airflow==3.0.4
asn1crypto==1.5.1
boto3==1.34.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cloudpickle==2.2.1
cryptography==43.0.1
filelock==3.16.1
idna==3.10
numpy==2.1.1
packaging==24.1
pandas==2.2.3
platformdirs==4.3.6
pyarrow==17.0.0
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
beautifulsoup4==4.12.2
six==1.16.0
snowflake-connector-python==3.17.1
sortedcontainers==2.4.0
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3 